Squashed commit of the following:
commit 4c2cf5b9e6b896b5fbf6898ff2552ba6de591fe8
Author: Mats Larsen <me@supergrecko.com>
Date:   Tue May 25 10:44:02 2021 +0200

    Undo separating samples into their own modules
