+++ llvm/samples/llvm/pom.xml
@@ -1,19 +1,32 @@
 <project>
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.bytedeco.llvm</groupId>
-
     <artifactId>samples</artifactId>
     <version>1.5.6-SNAPSHOT</version>
-    <modules>
-        <module>EmitBitcode</module>
-        <module>Factorial</module>
-        <module>OrcJIT</module>
-    </modules>
+    <properties>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+        <exec.mainClass>OrcJit</exec.mainClass>
+    </properties>
     <repositories>
         <repository>
             <id>sonatype-nexus-snapshots</id>
             <url>https://oss.sonatype.org/content/repositories/snapshots</url>
         </repository>
     </repositories>
-    <packaging>pom</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>org.bytedeco</groupId>
+            <artifactId>llvm-platform</artifactId>
+            <version>12.0.0-1.5.6-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.bytedeco</groupId>
+            <artifactId>libffi-platform</artifactId>
+            <version>3.3-1.5.6-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+    <build>
+        <sourceDirectory>.</sourceDirectory>
+    </build>
 </project>
